Cassandra文档添加性能测试说明

+++ demo-cassandra/demo-order-management-app/service/src/test/java/com/future/demo/ApplicationTests.java
@@ -42,7 +42,7 @@ public class ApplicationTests {
 
         List<OrderModel> orderModelList = this.orderMapper.selectAll();
 
-        BigDecimal orderId = orderModelList.get(0).getId();
+        Long orderId = orderModelList.get(0).getId();
 
         OrderDTO orderDTO = this.orderService.getById(orderId);
         Assertions.assertEquals(orderId, orderDTO.getId());
